Add files via upload
This is a basic project based on java which is linked to a database (Sql).i have used eclpse to create the java project with the required files for the different modules and functionalities used for Airline Management System.You can search a record by entering passenger id and you can search a booking by entering booking id also it is applicable in all the other functionalities(delete,insert and search).Similarly for other tables as well(Food_order,seat_confirmation)etc.  Also i have defined constraints in each table for normalization(preventing duplication of data).